WIB-5 Rename(viewmodels, models) | .spec.ts : 테스트 파일들 __test__로 이동
모든 파일의 수정사항은 단순히 파일 이동에 의한 import 수정
